feat: extract gateway_following package (service + schemas)
