+++ pyproject.toml
@@ -25,6 +25,7 @@ members = ["packages/*"]
 [tool.uv.sources]
 gateway_oss = { workspace = true }
 gateway_core = { workspace = true }
+gateway_following = { workspace = true }
 gateway_notes = { workspace = true }
 gateway_notes_rest = { workspace = true }
 gateway_notes_mcp = { workspace = true }
@@ -41,7 +42,7 @@ select = ["E", "F", "I", "UP"]
 ignore = ["E501"]
 
 [tool.ruff.lint.isort]
-known-first-party = ["gateway_core", "gateway_notes", "gateway_notes_rest", "gateway_notes_mcp", "gateway_oss", "gateway_rest_common", "gateway_mcp_common", "substack_gateway"]
+known-first-party = ["gateway_core", "gateway_following", "gateway_notes", "gateway_notes_rest", "gateway_notes_mcp", "gateway_oss", "gateway_rest_common", "gateway_mcp_common", "substack_gateway"]
 
 [tool.ruff.format]
 quote-style = "double"
@@ -58,6 +59,7 @@ default_format = "pretty"
 testpaths = [
   "packages/gateway_oss/tests",
   "packages/gateway_core/tests",
+  "packages/gateway_following/tests",
   "packages/gateway_notes/tests",
 ]
 addopts = "--import-mode=importlib"
